# app/Models/Product.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Product extends Model
{
    use HasFactory;

    protected $fillable = [
        'name', 'description', 
        'price', 'stock'
    ]; 

    protected $casts = [
        'price' => 'decimal:2', 
        'stock' => 'integer', 
    ];

    public function orderItems()
    {
        return $this->hasMany(OrderItem::class);
    }
}